move portfolio to pages folder, rename vars
