+++ src/components/shell/bank/BankAdd.test.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { render, screen } from '@testing-library/react';
-import BankSearch from './BankAdd';
+import BankAdd from './BankAdd';
 
 describe('BankSearch', () => {
   test('Search field exists', () => {
-    render(<BankSearch />);
+    render(<BankAdd />);
 
     expect(screen.getByRole('textbox')).toBeInTheDocument();
   });
